# docs/openapi/famlin.yaml
...
openapi: 3.0.3
info:
  title: Famlin API
  version: current
  description: |
    Every Famlin server exposes this REST API under `/api` — the same API the
    official mobile and web apps use. With a [personal access token](/developers/api)
    you can build your own integrations against your family's server.

    All content is scoped by **group membership**: you can only read and write
    content in groups (families) you belong to, and there is no way around that
    per token. See [Using the API](/developers/api) for authentication,
    conventions, and worked examples.
servers:
  - url: https://famlin.example.com
    description: Your own Famlin deployment
security:
  - bearerAuth: []
...
paths:
  /api/auth/login:
    post:
      operationId: login
      tags: [Account]
      summary: Log in with email and password
      security: []
      description: |
        Returns a session token (JWT) valid for 30 days — the credential the
        official apps use. For long-lived integrations prefer a
        [personal access token](/developers/api#personal-access-tokens-recommended):
        session tokens expire and are all invalidated when you change your
        password.

        Rate-limited to 10 attempts per 15 minutes per IP.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                inviteToken:
                  type: string
                  description: |
                    Optional invite token; when valid (and matching this
                    account's email, if the invite is email-restricted) the
                    account joins the invite's group as part of the login.
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: Session JWT, valid 30 days.
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/auth/me:
    get:
      operationId: getMe
      tags: [Account]
      summary: Get the authenticated user
      responses:
        '200':
          description: The current user, including their groups
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/User'
                  - type: object
                    properties:
                      groups:
                        type: array
                        items:
                          $ref: '#/components/schemas/Group'
        '401':
          $ref: '#/components/responses/Unauthorized'
    patch:
      operationId: updateMe
      tags: [Account]
      summary: Update profile and notification preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                avatarUrl:
                  type: string
                  nullable: true
                  description: |
                    An `/uploads/...` path previously returned by the upload
                    endpoint. Arbitrary external URLs are rejected. `null`
                    clears the avatar.
                  example: /uploads/8d0f3c9a-1b2e-4f5a-9c8d-7e6f5a4b3c2d.jpg
                emailOnNewPost: { type: boolean }
                emailOnNewComment: { type: boolean }
                emailOnNewLike: { type: boolean }
                pushOnNewPost: { type: boolean }
                pushOnNewComment: { type: boolean }
                pushOnNewLike: { type: boolean }
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/change-password:
    post:
      operationId: changePassword
      tags: [Account]
      summary: Change your password
      description: |
        Immediately invalidates every existing session token for the account.
        Personal access tokens deliberately stay valid — revoke those
        individually if needed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string }
                newPassword:
                  type: string
                  minLength: 8
                  maxLength: 100
      responses:
        '200':
          description: Password changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        A personal access token (`famlin_pat_...`, created in the web app) or
        a session JWT from `POST /api/auth/login`. Both are sent the same way:
        `Authorization: Bearer <token>`.
...
  responses:
    Unauthorized:
      description: Missing, invalid, expired, or revoked credential
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    ValidationError:
      description: Invalid request body or query (see `details`)
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  details:
                    type: object
                    description: Zod field errors.
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Human-readable, translated via the `Accept-Language` header (`en`/`nl`).
      required: [error]
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true
...
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string, format: email }
        name: { type: string }
        avatarUrl:
          type: string
          nullable: true
        isAdmin: { type: boolean }
        emailOnNewPost: { type: boolean }
        emailOnNewComment: { type: boolean }
        emailOnNewLike: { type: boolean }
        pushOnNewPost: { type: boolean }
        pushOnNewComment: { type: boolean }
        pushOnNewLike: { type: boolean }
    Group:
      type: object
      properties:
        id: { type: string }
        name: { type: string, example: Family Smith }
        description:
          type: string
          nullable: true
        createdAt: { type: string, format: date-time }
        joinedAt:
          type: string
          format: date-time
          description: When *you* joined this group.
...